ログインの不具合を修正

+++ src/config/app-features.config.ts
@@ -14,7 +14,7 @@ export const APP_FEATURES = {
     /** 本登録ページに進捗インディケーターを表示する */
     showRegistrationSteps: true,
     /** パスワード入力モード: "single"=確認なし, "double"=確認あり */
-    passwordInputMode: "double" as "single" | "double",
+    passwordInputMode: "single" as "single" | "double",
   },
   admin: {
     appearance: {
